feat(drive-fsa): implement move folder

+++ packages/drive-fsa/composables/moveEntry.ts
@@ -1,28 +1,48 @@
 import { destroyEntry } from "./destroyEntry"
 import { findEntry } from "./findEntry"
+import { listEntries } from "./listEntries"
+import { makeDirectoryEntry } from "./makeDirectoryEntry"
 import { readEntry } from "./readEntry"
 import { writeEntry } from "./writeEntry"
 
 export async function moveEntry(rootHandle: FileSystemDirectoryHandle, from: string, to: string){
     const entry = await findEntry(rootHandle, from)
 
+	const toExists = await findEntry(rootHandle, to)
+
+	if (toExists) {
+		throw new Error('Target already exists')
+	}
+
     if (!entry) {
         throw new Error('File not found')
     }
 
+
     if (entry.handle instanceof FileSystemFileHandle) {
+		const contents = await readEntry(rootHandle, from)
 
-        const contents = await readEntry(rootHandle, from)
-    
-        await writeEntry(rootHandle, to, contents)
-    
-        await destroyEntry(rootHandle, from)
+		await writeEntry(rootHandle, to, contents)
 
-        return
-    }
+		await destroyEntry(rootHandle, from)
 
-    if (entry.handle instanceof FileSystemDirectoryHandle) {
-        throw new Error('Folders move not implemented yet')
-    }
+		return
+	}
+
+	if (entry.handle instanceof FileSystemDirectoryHandle) {
+		const entries = await listEntries(rootHandle, entry.path)
+
+		await makeDirectoryEntry(rootHandle, to)
+
+		for await (const e of entries) {
+			const eFrom = e.path
+			const eTo = e.path.replace(from, to)
+
+			await moveEntry(rootHandle, eFrom, eTo)
+		}
+
+		await destroyEntry(rootHandle, from) 
+
+	}
 
 }
